pyproject.toml
[project]
name = "botler"
version = "0.1.0"
description = "An agent runtime you own"
requires-python = ">=3.14"
dependencies = [
    "pydantic>=2.0",
    "pydantic-ai>=0.1",
    "anthropic>=0.30",
    "click>=8.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
botler = "botler.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/botler"]

[tool.ruff]
line-length = 100
